move register sample gql into folder
