--- components/pages/submission-system/program-sample-registration/CLEAR_CLINICAL_REGISTRATION_MUTATION.gql
@@ -1,3 +1,0 @@
-mutation CLEAR_CLINICAL_REGISTRATION_MUTATION($shortName: String!, $registrationId: String!) {
-  clearClinicalRegistration(shortName: $shortName, registrationId: $registrationId)
-}
