--- components/pages/submission-system/program-sample-registration/GET_REGISTRATION.gql
@@ -1,7 +1,0 @@
-#import "./REGISTRATION_FRAGMENT.gql"
-
-query($shortName: String!) {
-  clinicalRegistration(shortName: $shortName) {
-    ...Registration
-  }
-}
