--- components/pages/submission-system/program-sample-registration/REGISTRATION_FRAGMENT.gql
@@ -1,40 +1,0 @@
-fragment Registration on ClinicalRegistrationData {
-  id
-  programShortName
-  creator
-  fileName
-  createdAt
-  records {
-    row
-    fields {
-      name
-      value
-    }
-  }
-  errors {
-    type
-    message
-    row
-    field
-    value
-    sampleId
-    donorId
-    specimenId
-  }
-  newDonors {
-    count
-    rows
-  }
-  newSpecimens {
-    count
-    rows
-  }
-  newSamples {
-    count
-    rows
-  }
-  alreadyRegistered {
-    count
-    rows
-  }
-}
