--- components/pages/submission-system/program-sample-registration/UPLOAD_REGISTRATION.gql
@@ -1,14 +1,0 @@
-#import "./REGISTRATION_FRAGMENT.gql"
-
-mutation UPLOAD_REGISTRATION($shortName: String!, $registrationFile: Upload!) {
-  uploadClinicalRegistration(shortName: $shortName, registrationFile: $registrationFile) {
-    ... on ClinicalRegistrationData {
-      ...Registration
-    }
-
-    ... on ClinicalRegistrationInvalid {
-      error
-      code
-    }
-  }
-}
