+++ components/pages/submission-system/program-sample-registration/index.tsx
@@ -13,10 +13,10 @@ import Progress, { PROGRESS_STATUS } from 'uikit/Progress';
 import TitleBar from 'uikit/TitleBar';
 import Typography from 'uikit/Typography';
 import SubmissionLayout from '../layout';
-import CLEAR_CLINICAL_REGISTRATION_MUTATION from './CLEAR_CLINICAL_REGISTRATION_MUTATION.gql';
+import CLEAR_CLINICAL_REGISTRATION_MUTATION from './gql/CLEAR_CLINICAL_REGISTRATION_MUTATION.gql';
 import FileTable from './FileTable';
 import NoDataMessage from './FileTable/NoDataMessage';
-import GET_REGISTRATION from './GET_REGISTRATION.gql';
+import GET_REGISTRATION from './gql/GET_REGISTRATION.gql';
 import Instructions from './Instructions';
 import { FileEntry } from './FileTable';
 import { ERROR_CODES } from './common';
